requirements/local.txt
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-model-utils==4.5.0
idna==3.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
pyproject-toml==0.0.10
python3-openid==3.2.0
pytz==2024.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
setuptools==69.5.1
sqlparse==0.5.0
toml==0.10.2
tzdata==2024.1
Unipath==1.1
urllib3==2.2.1
wheel==0.43.0